limitovany decimal input na 1 misto

+++ app/src/main/java/dev/skaba/soma/app/ui/features/food/viewmodel/FoodFormState.kt
@@ -2,7 +2,7 @@ package dev.skaba.soma.app.ui.features.food.viewmodel
 
 data class FoodFormState(
   val name: String = "",
-  val kcal: Int? = null,
+  val kcal: Float? = null,
   val carbs: Float? = null,
   val proteins: Float? = null,
   val fats: Float? = null,
